# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API for user to post something",
        "title": "MyGram API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hamid Baehaqi",
            "email": "hamid1bae1@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/social-media": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "list all social media from logged in user, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social-media"
                ],
                "summary": "get social media from logged in user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/socialmedia.SocialMedia"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "add social media to user, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social-media"
                ],
                "summary": "add social media to user",
                "parameters": [
                    {
                        "description": "social media info",
                        "name": "socialmedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/socialmedia.AddSocialMediaInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/socialmedia.SocialMedia"
                        }
                    }
                }
            }
        },
        "/v1/social-media/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "update social media, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social-media"
                ],
                "summary": "update social media",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "social media id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "social media info",
                        "name": "socialmedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/socialmedia.UpdateSocialMediaInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/socialmedia.SocialMedia"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete social media, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "social-media"
                ],
                "summary": "delete social media",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "social media id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/v1/user": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "update user, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "update user",
                "parameters": [
                    {
                        "description": "user info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserUpdateDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.UserUpdateResponseDto"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "delete user, auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "delete user",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/v1/user/{id}": {
            "get": {
                "description": "get user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "get user by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserWithSocialMediaDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "socialmedia.AddSocialMediaInput": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "instagram"
                },
                "url": {
                    "type": "string",
                    "example": "url"
                }
            }
        },
        "socialmedia.SocialMedia": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "socialmedia.UpdateSocialMediaInput": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string",
                    "example": "url"
                }
            }
        },
...
        "user.UserUpdateDto": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "luigi@mail.com"
                },
                "username": {
                    "type": "string",
                    "example": "luigi"
                }
            }
        },
        "user.UserUpdateResponseDto": {
            "type": "object",
            "properties": {
                "token_id": {
                    "type": "string"
                }
            }
        },
        "user.UserWithSocialMediaDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "social_medias": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/socialmedia.SocialMedia"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}